updated bg

+++ commands/bg.c
@@ -22,13 +22,20 @@ void command_bg(token_mat arg){
     }
 
     int job_pid = run[job_number-1]->pid;
+    char pid[MAX_TOKENS];
+    sprintf(pid,"%d",job_pid);
+    char* stat = status(pid);
 
+    if(strcmp(stat,"R") == 0 || strcmp(stat,"S") == 0){
+        printf("The background process is already running!\n");
+        return;
+    }
     if(kill(job_pid, SIGCONT) < 0){
         perror("Error SIGCONT! ");
         return;
     }
 
-    printf("[%d] %s %d stopped background process resumed in background!\n", job_number, run[job_number-1]->name, job_pid);
+    printf("[%d] %s [%d] stopped background process resumed in background!\n", job_number, run[job_number-1]->name, job_pid);
 
 
 }
